Move text editor functions to Doc
